Slightly reduce the space complexity.

+++ NP-complete/tsp/src/test_tsp.h
@@ -36,7 +36,7 @@ int testTSPDP() {
       xy {{0, 0}, {-2, -7}, {4, -7}, {0, -5}};
 
   double min_dist = 21.8907;
-  std::vector<size_t> tsp_path {0, 2, 1, 3, 0};
+  std::vector<unsigned int> tsp_path {0, 2, 1, 3, 0};
 
   auto result = tspDP(xy);
